requirements.txt
altair==5.0.1
APScheduler==3.10.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
configparser==5.3.0
confluent-kafka==2.1.1
contourpy==1.1.0
cycler==0.11.0
db-dtypes==1.1.1
debugpy==1.6.7
decorator==5.1.1
dnspython==2.3.0
executing==1.2.0
finnhub-python==2.4.18
fonttools==4.40.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.19.1
google-cloud-bigquery==3.11.1
google-cloud-core==2.3.2
google-cloud-secret-manager==2.16.1
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.3
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkg_resources==0.0.0
platformdirs==3.5.3
plotly==5.15.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==4.23.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.15.1
pymongo==4.3.3
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.0
requests==2.31.0
rich==13.4.1
rsa==4.9
schedule==1.2.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit-autorefresh==0.0.1
streamlit-nightly==1.23.2.dev20230614
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.16
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
zipp==3.15.0
